[#176] Add an action to flush a handle
Problem: during logging one often wants data to reach destination
immediately, but logging actions provided out of the box do not
flush handles they log to.

Solution: we add a new action to `co-log-core`. This action does
not _write_ anything, but flushes the handler. We call it
`logFlush` for consistency with other `log*` actions.
It can be used as `logPrintStderr <> logFlush stderr`.